Completely refactored into a multi-crate workspace.
