+++ Cargo.toml
@@ -1,49 +1,7 @@
-[package]
-name = "nomen"
-version = "0.2.0"
-edition = "2021"
-build = "build.rs"
-rust-version = "1.71"
-default-run = "nomen"
+[workspace]
 
-[[bin]]
-name = "nomen"
-
-[[bin]]
-name = "nomen-cli"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.69"
-askama = {version = "0.12.0", features = ["with-axum", "serde-json"]}
-askama_axum = "0.3.0"
-axum = {version = "0.6.11"}
-axum-extra = "0.7.4"
-bitcoin = {version = "0.30.0", features = ["base64", "serde"] }
-bitcoin_hashes = { version = "0.12.0", features = ["serde"] }
-bitcoincore-rpc = "0.17.0"
-clap = { version = "4.1.8", features = ["derive"] }
-derive_more = "0.99.17"
-elegant-departure = { version = "0.2.1", features = ["tokio"] }
-env_logger = "0.10.0"
-hex = "0.4.3"
-itertools = "0.10.5"
-log = "0.4.17"
-nostr-sdk = "0.24.0"
-rand = "0.8.5"
-regex = "1.7.1"
-ripemd = "0.1.3"
-secp256k1 = { version = "0.27.0", features = ["rand-std"] }
-serde = { version = "1.0.152", features = ["derive"] }
-serde-hex = "0.1.0"
-serde_json = "1.0.94"
-sha2 = "0.10.6"
-sqlx = { version = "0.6.2", features = ["runtime-tokio-rustls", "sqlite"] }
-time = { version = "0.3.20", features = ["formatting", "macros"] }
-tokio = { version = "1.26.0", features = ["full"] }
-toml = "0.7.2"
-yansi = "0.5.1"
-
-[build-dependencies]
-vergen = { version = "8.0.0", features = ["build", "git", "gitcl"] }
+members = [
+  "nomen_core",
+  "nomen",
+  "nomen-cli"
+]
