+++ nomen-cli/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "nomen-cli"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+nomen_core = { path =  "../nomen_core" }
+anyhow = "1.0.75"
+clap = "4.4.4"
+nostr-sdk = "0.24.0"
+secp256k1 = { version = "0.27.0", features = ["rand-std", "bitcoin-hashes"] }
+hex = { version = "0.4.3", features = ["serde"] }
