+++ nomen/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "nomen"
+version = "0.1.0"
+edition = "2021"
+build = "build.rs"
+default-run = "nomen"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.75"
+askama = {version = "0.12.0", features = ["with-axum", "serde-json"]}
+askama_axum = "0.3.0"
+axum = {version = "0.6.11"}
+axum-extra = "0.7.4"
+clap = { version = "4.1.8", features = ["derive"] }
+env_logger = "0.10.0"
+hex = { version = "0.4.3", features = ["serde"] }
+log = "0.4.20"
+nomen_core = { path = "../nomen_core" }
+nostr-sdk = "0.24.0"
+rand = { version = "0.8.5", features = ["serde"] }
+secp256k1 = { version = "0.27.0", features = ["rand-std", "bitcoin-hashes"] }
+serde = { version = "1.0.188", features = ["derive"] }
+serde_json = "1.0.107"
+tokio = { version = "1.32.0", features = ["full"] }
+toml = "0.8.0"
+yansi = "0.5.1"
+sqlx = { version = "0.6.2", features = ["runtime-tokio-rustls", "sqlite"] }
+bitcoin = { version = "0.30.1", features = ["base64", "rand", "serde"] }
+elegant-departure = { version = "0.2.1", features = ["tokio"] }
+itertools = "0.11.0"
+bitcoincore-rpc = "0.17.0"
+
+
+[build-dependencies]
+vergen = { version = "8.0.0", features = ["build", "git", "gitcl"] }
