+++ nomen/src/db/queries/insert_blockchain_index.sql
