+++ nomen/src/db/queries/insert_name_event.sql
