+++ nomen/src/db/queries/insert_namespace.sql
