+++ nomen/src/subcommands/index/REFACTOR.md
