+++ nomen/src/subcommands/index/events/mod.rs
