+++ nomen_core/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "nomen_core"
+version = "0.2.0"
+edition = "2021"
+rust-version = "1.71"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.69"
+bitcoin = {version = "0.30.0", features = ["base64", "serde"] }
+bitcoin_hashes = { version = "0.12.0", features = ["serde"] }
+derive_more = "0.99.17"
+hex = { version = "0.4.3", features = ["serde"] }
+itertools = "0.10.5"
+log = "0.4.17"
+nostr-sdk = "0.24.0"
+rand = "0.8.5"
+regex = "1.7.1"
+ripemd = "0.1.3"
+secp256k1 = { version = "0.27.0", features = ["rand-std"] }
+serde = { version = "1.0.152", features = ["derive"] }
+serde-hex = "0.1.0"
+serde_json = "1.0.94"
+sha2 = "0.10.6"
+time = { version = "0.3.20", features = ["formatting", "macros"] }
