+++ nomen_core/src/lib.rs
@@ -1,0 +1,3 @@
+#![allow(unused)]
+
+pub mod util;
