+++ nomen_core/src/util/kind.rs
