+++ nomen_core/src/util/mod.rs
@@ -6,7 +6,6 @@ mod name;
 mod nostr;
 mod nsid;
 mod nsid_builder;
-mod transfer;
 
 use anyhow::bail;
 pub use extractor::*;
@@ -18,13 +17,6 @@ pub use nostr::*;
 pub use nsid::*;
 pub use nsid_builder::*;
 use time::{macros::format_description, OffsetDateTime};
-pub use transfer::*;
-use yansi::Paint;
-
-use crate::{
-    config::{Cli, Config},
-    db,
-};
 
 pub enum NameKind {
     Name = 38300,
@@ -46,19 +38,6 @@ impl TryFrom<nostr_sdk::Kind> for NameKind {
         };
         Ok(nk)
     }
-}
-
-pub fn tag_print(tag: &str, message: &str) {
-    println!("{}: {}", Paint::green(tag), message);
-}
-
-pub async fn check_name_availability(config: &Config, name: &str) -> anyhow::Result<()> {
-    let conn = config.sqlite().await?;
-    let available = db::name_available(&conn, name).await?;
-    if !available {
-        bail!("Name {name} already exists");
-    }
-    Ok(())
 }
 
 pub fn format_time(timestamp: i64) -> anyhow::Result<String> {
