+++ nomen_core/src/util/nsid.rs
