+++ nomen_core/src/util/nsid_builder.rs
