--- src/lib.rs
@@ -1,6 +1,0 @@
-#![allow(unused)]
-
-pub mod config;
-pub mod db;
-pub mod subcommands;
-pub mod util;
